new changes on tartaglias implementation
