+++ imports.go
@@ -8,7 +8,6 @@ import (
 	_ "github.com/genshinsim/gsim/internal/characters/ayaka"
 	_ "github.com/genshinsim/gsim/internal/characters/beidou"
 	_ "github.com/genshinsim/gsim/internal/characters/bennett"
-	_ "github.com/genshinsim/gsim/internal/characters/childe"
 	_ "github.com/genshinsim/gsim/internal/characters/chongyun"
 	_ "github.com/genshinsim/gsim/internal/characters/diluc"
 	_ "github.com/genshinsim/gsim/internal/characters/diona"
@@ -27,6 +26,7 @@ import (
 	_ "github.com/genshinsim/gsim/internal/characters/raiden"
 	_ "github.com/genshinsim/gsim/internal/characters/rosaria"
 	_ "github.com/genshinsim/gsim/internal/characters/sucrose"
+	_ "github.com/genshinsim/gsim/internal/characters/tartaglia"
 	_ "github.com/genshinsim/gsim/internal/characters/venti"
 	_ "github.com/genshinsim/gsim/internal/characters/xiangling"
 	_ "github.com/genshinsim/gsim/internal/characters/xiao"
